split /scores into /scores (takes in positions, outputs risk score) and /balances (returns positions)
