+++ api/scores/get.py
@@ -5,97 +5,12 @@ import numpy as np
 import json
 import math
 
-def verify_account(params):
-    if "account" not in params:
-        raise InputException("missing account") # check policy id exists
-    try:
-        addr = w3.toChecksumAddress(params["account"])
-        if not w3.isAddress(addr):
-            raise "invalid address"
-        return addr
-    except Exception as e:
-        raise InputException(f"invalid account: '{params['account']}'") # check policy id exists
-
-def verify_params(params):
-    if params is None:
-        raise InputException("missing params")
-    return {
-        "account": verify_account(params)
-    }
-
-def fetch_eth_price():
-    try:
-        url = f"https://api.zapper.fi/v1/prices/0x0000000000000000000000000000000000000000?network=ethereum&timeFrame=hour&currency=USD&api_key=96e0cc51-a62e-42ca-acee-910ea7d2a241"
-        response = requests.get(url, timeout=600)
-        response.raise_for_status()
-        res = response.json()
-        # gives prices over the last hour, average
-        prices = res["prices"]
-        count = 0
-        s = 0
-        for price in prices:
-            count += 1
-            s += price[1]
-        price = s / count
-        if price <= 1000 or price >= 10000:
-            throw("price out of range")
-        return price
-    except Exception as e:
-        print(e)
-        raise Exception("error fetching data")
-
-def fetch_positions(params):
-    try:
-        api_key = "96e0cc51-a62e-42ca-acee-910ea7d2a241" # only key key, public
-        url = f"https://api.zapper.fi/v1/balances?api_key={api_key}&addresses[]={params['account']}"
-        response = requests.get(url, timeout=600)
-        response.raise_for_status()
-        return response.text
-    except Exception as e:
-        print(e)
-        raise Exception("error fetching data")
-
-def parse_positions(s):
-    try:
-        positions = []
-        while True:
-            index = s.find('{')
-            index2 = s.find('}\n')
-            if index == -1 or index2 == -1:
-                break
-            position = json.loads(s[index:index2+1])
-            positions.append(position)
-            s = s[index2+1:]
-        # order by network, app
-        positions = list(sorted(positions, key = lambda pos: f"{pos['network']} {pos['appId']}"))
-        return positions
-    except Exception as e:
-        raise Exception("error parsing data")
-
-def clean_positions(positions2, account):
-    eth_price = fetch_eth_price()
-    try:
-        positions3 = []
-        account = account.lower()
-        for position in positions2:
-            # filter out zero balance positions
-            if len(position["balances"][account]["products"]) == 0:
-                continue
-            # filter out nfts and tokens
-            if position["appId"] == "tokens" or position["appId"] == "nft":
-                continue
-            # flatten
-            balanceUSD = 0
-            for pos in position["balances"][account]["products"]:
-                for asset in pos["assets"]:
-                    balanceUSD += asset["balanceUSD"]
-            position["balanceUSD"] = balanceUSD
-            position["balanceETH"] = balanceUSD / eth_price
-            position.pop("balances", None)
-            positions3.append(position)
-        return positions3
-    except Exception as e:
-        raise Exception("error parsing data")
+def verify_positions(positions_in):
+    # TODO: input sanitization
+    positions_out = json.loads(positions_in)
+    account = positions_out['account']
+    positions_out = positions_out['positions']
+    return account, positions_out
 
 def calculate_weights(positions):
     try:
@@ -108,28 +23,19 @@ def calculate_weights(positions):
     except Exception as e:
         raise Exception(f"error calculating weights. Error: {e}")
 
-def get_scores(params):
-    params2 = verify_params(params)
-    positions = fetch_positions(params2)
-    positions2 = parse_positions(positions)
-    positions3 = clean_positions(positions2, params2["account"])
-    positions4 = calculate_weights(positions3)
+def get_scores(event):
+    account, positions = verify_positions(event["body"])
+    positions4 = calculate_weights(positions)
     # TODO: Lock in on field names for appId as 'address' isn't ideal atm
-    if len(positions3) == 0:
-        return {'address': params2["account"], 'protocols': positions3}
-    for i in positions3:
-        i['address'] = i['appId']
-        i['symbol'] = i['network']
-        del i['appId']
-        del i['network']
-    
+    if len(positions) == 0:
+        return json.dumps({'address': account, 'protocols': []})
 
     # local functions
     #
     def get_protocol(accountIn, portfolioInput, protocolMapInput, rateTableInput):
         # Left join protocols in account with known protocol tiers then remove nans
         scored_portfolio = pd.merge(portfolioInput,protocolMapInput, left_on='protocol', right_on='protocol',how='left')
-        #scored_portfolio['tier'] = scored_portfolio['tier'].replace(np.nan, 'tier4')  # sets unknown protocol to highest risk tier 
+        #scored_portfolio['tier'] = scored_portfolio['tier'].replace(np.nan, 'tier4')  # sets unknown protocol to highest risk tier
         # log unknown protocols to S3
         #inds = pd.isnull(scored_portfolio['tier']).to_numpy().nonzero()
         indexInScoringDb = list(scored_portfolio.loc[pd.isna(scored_portfolio["tier"]), :].index)
@@ -139,10 +45,10 @@ def get_scores(params):
             outputUnknownProtocols = unknownProtocols.to_list()
             date_string = datetime.now(). strftime("%Y-%m-%d")
             s3_put("to-be-scored/"+date_string+"/"+accountIn+".json", json.dumps(outputUnknownProtocols))
-        
+
         # uncomment for lambda
         #s3_put("to-be-scored/"+date_string+"/"+accountIn+".json", json.dumps(outputUnknownProtocols))
-        scored_portfolio['tier'] = scored_portfolio['tier'].replace(np.nan, 0)  # sets unknown protocol to highest risk tier 
+        scored_portfolio['tier'] = scored_portfolio['tier'].replace(np.nan, 0)  # sets unknown protocol to highest risk tier
 
         # aggregate portfolio positions by tier
         combinedTable = pd.merge(scored_portfolio,rateTableInput, left_on='tier', right_on='tier',how='left')
@@ -172,8 +78,8 @@ def get_scores(params):
         return combinedTable4
     #
     def create_matrix(correlValue,countCategory,index):
-        
-        if countCategory!=0:                
+
+        if countCategory!=0:
             matrix=np.zeros(shape=(countCategory,countCategory))
             for j in range(countCategory):
                 for k in range(countCategory):
@@ -182,7 +88,7 @@ def get_scores(params):
                     else:
                         matrix[k][j] = correlValue['correlation'][index]
                         matrix[j][k] = correlValue['correlation'][index]
-            
+
             return matrix
     #
     def create_protocol_correlation(portfolioMap,correlValue):
@@ -202,7 +108,7 @@ def get_scores(params):
         #print('correlCatIn ',correlCatIn)
         for i in portfolioMap['category'].unique():
             if i in correlCatIn:
-                categoryarray.append(correlCatIn[i]) 
+                categoryarray.append(correlCatIn[i])
         correlCatIn_values=np.array(categoryarray)
         #print('correlCatIn_values ',correlCatIn_values)
         ra_line=np.array([ra_column])
@@ -210,14 +116,14 @@ def get_scores(params):
         #print('this is ra_column2',ra_column2)
         #print('this is correlCatIn_values[0:N,0:N]',correlCatIn_values)
         new_matrix=np.matmul(correlCatIn_values[0:N,0:N],ra_column2)
-        portfolioRateScore=math.sqrt(np.matmul(ra_line,new_matrix))       
+        portfolioRateScore=math.sqrt(np.matmul(ra_line,new_matrix))
         return portfolioRateScore
 
     def rate_engine(accountIn, positions):
         # Log the positions by account
         date_string = datetime.now(). strftime("%Y-%m-%d")
         s3_put("asked-for-quote/"+date_string+"/"+accountIn+".json", json.dumps(positions))
-        
+
         # Get the published rate data and create dataframes
         correlCat_file = s3_get("current-rate-data/correlCat.json")
         correlCatJson_object = json.loads(correlCat_file)
@@ -231,19 +137,19 @@ def get_scores(params):
         corrValueFile = s3_get("current-rate-data/corrValue.json") # will be from S3 not local
         corrValueJson_object = json.loads(corrValueFile)
         correlValue=pd.DataFrame(corrValueJson_object)
-        
+
         rateTableJson_file = s3_get("current-rate-data/rateTable.json")
         rateTableJson_object = json.loads(rateTableJson_file)
         rateTable = pd.DataFrame(rateTableJson_object)
 
         # Init a dataframe to store an account's positions
-        
+
         portfolio = pd.DataFrame(positions)
         portfolio.columns =['balanceUSD', 'balance','protocol', 'network']
 
         #Portfolio of protocols
         balanceByTier = get_protocol(accountIn, portfolio, protocolMap, rateTable)
-        balanceByTier['category'] = balanceByTier['category'].replace(np.nan, 'unrated')  
+        balanceByTier['category'] = balanceByTier['category'].replace(np.nan, 'unrated')
 
         #print(balanceByTier)
 
@@ -274,19 +180,19 @@ def get_scores(params):
         TotalRate=RpTotal+TotalScore
         Rate_percentage=TotalRate/BalanceTotal
 
-        # Table data 
-        balanceByTier['json'] = balanceByTier.to_json(orient='records', lines=True).splitlines()       
+        # Table data
+        balanceByTier['json'] = balanceByTier.to_json(orient='records', lines=True).splitlines()
         rateOut = {'address': accountIn,'addressRP':TotalRate,'currentRate':Rate_percentage,'protocols': list(map(json.loads, balanceByTier['json'].tolist()))}
         return rateOut
 
-    
-    rateCard = rate_engine(params["account"],positions3) 
+
+    rateCard = rate_engine(account, positions)
     #print(rateCard)
-    return json.dumps(rateCard, indent=2) 
+    return json.dumps(rateCard, indent=2)
 
 def handler(event, context):
     try:
-        response_body = get_scores(event["queryStringParameters"])
+        response_body = get_scores(event)
         return {
             "statusCode": 200,
             "body": response_body,
